refactor: split temper verify tests out of the command module
Keeps the production file under the 500-line ratchet after the
multi-dir composite proof landed.
